moving css styles around, mainly
moved the light mode button css styles to a separate file

Added comments regarding WCAG compliance and style guide decisions.  For consistency, none of the custom colors will be used with text (some are fine with WCAG contrast, but others are not.  We made the decision to not use any with text for consistency)